+++ pages/_app.js
@@ -5,7 +5,7 @@ import { FadeIn } from '@carbonplan/components'
 import '@carbonplan/components/fonts.css'
 import '@carbonplan/components/globals.css'
 import theme from '@carbonplan/theme'
-import useStore from '../components/store'
+import useStore from '../components/use-store'
 import Header from '../components/header'
 import Layout from '../components/layout'
 import Info from '../components/info'
